Updated autobidding functionality

+++ WebContent/autobidPlaceProcess.jsp
@@ -29,29 +29,33 @@
     double bidVal = Double.parseDouble(request.getParameter("bidValue"));
     double bidIncr = Double.parseDouble(request.getParameter("increment"));
     double bidCeil = Double.parseDouble(request.getParameter("ceiling"));
-    Connection conn = new Database().getConnection();
-    try {
-        if (conn.createStatement().executeQuery("SELECT listingID FROM auctionItem WHERE listingID = " + listingID + " AND closingDate > NOW();").next()) {
-            PreparedStatement pst = conn.prepareStatement("INSERT INTO bid (listingID, bidder, bidValue, bidDate) VALUES (?,?,?,NOW());");
-            pst.setInt(1, listingID);
-            pst.setInt(2, acc.getAccountNumber());
-            pst.setDouble(3, bidVal);
-            pst.executeUpdate();
-            // flush any old autobid entries for this user and listing
-            conn.createStatement().executeUpdate("DELETE FROM autobid WHERE listingID = " + listingID + " AND userID = " + acc.getAccountNumber() + ";");
-            pst = conn.prepareStatement("INSERT INTO autobid (listingID, userID, ceiling, increment) VALUES (?,?,?,?);");
-            pst.setInt(1, listingID);
-            pst.setInt(2, acc.getAccountNumber());
-            pst.setDouble(3, bidCeil);
-            pst.setDouble(4, bidIncr);
-            pst.executeUpdate();
-            CheckAutoBids.run();
-            out.println("<div class=\"container signin\"><p>Successfully placed bid!<br><br><a href=\"auctionList.jsp\">Return to listings</a>.</p> </div>");
-        } else {
-            out.println("<div class=\"container signin\"><p>Sorry, the auction has closed<br><br><a href=\"auctionList.jsp\">Return to listings</a>.</p> </div>");
+    if (bidCeil < bidVal) {
+        out.println("<div class=\"container signin\"><p>Your bid ceiling cannot be less than your bid. Please go back and edit your auto-bid<br><br><a href=\"auctionList.jsp\">Return to listings</a>.</p> </div>");
+    } else {
+        Connection conn = new Database().getConnection();
+        try {
+            if (conn.createStatement().executeQuery("SELECT listingID FROM auctionItem WHERE listingID = " + listingID + " AND closingDate > NOW();").next()) {
+                PreparedStatement pst = conn.prepareStatement("INSERT INTO bid (listingID, bidder, bidValue, bidDate) VALUES (?,?,?,NOW());");
+                pst.setInt(1, listingID);
+                pst.setInt(2, acc.getAccountNumber());
+                pst.setDouble(3, bidVal);
+                pst.executeUpdate();
+                // flush any old autobid entries for this user and listing
+                conn.createStatement().executeUpdate("DELETE FROM autobid WHERE listingID = " + listingID + " AND userID = " + acc.getAccountNumber() + ";");
+                pst = conn.prepareStatement("INSERT INTO autobid (listingID, userID, ceiling, increment) VALUES (?,?,?,?);");
+                pst.setInt(1, listingID);
+                pst.setInt(2, acc.getAccountNumber());
+                pst.setDouble(3, bidCeil);
+                pst.setDouble(4, bidIncr);
+                pst.executeUpdate();
+                CheckAutoBids.run();
+                out.println("<div class=\"container signin\"><p>Successfully placed bid!<br><br><a href=\"auctionList.jsp\">Return to listings</a>.</p> </div>");
+            } else {
+                out.println("<div class=\"container signin\"><p>Sorry, the auction has closed<br><br><a href=\"auctionList.jsp\">Return to listings</a>.</p> </div>");
+            }
+        } catch (SQLException throwables) {
+            throwables.printStackTrace();
         }
-    } catch (SQLException throwables) {
-        throwables.printStackTrace();
     }
